Move everything out of ssl protobuf reader
