+++ src/software/proto/message_translation/ssl_wrapper_test.cpp
@@ -2,14 +2,14 @@
 
 #include <gtest/gtest.h>
 
-TEST(SSLWrapperMessageTranslatorTest, test_create_empty_message)
+TEST(SSLWrapperTest, test_create_empty_message)
 {
     auto wrapper_packet = createWrapperPacket(nullptr, nullptr);
     EXPECT_FALSE(wrapper_packet->has_geometry());
     EXPECT_FALSE(wrapper_packet->has_detection());
 }
 
-TEST(SSLWrapperMessageTranslatorTest, test_create_wrapper_only_with_geometry)
+TEST(SSLWrapperTest, test_create_wrapper_only_with_geometry)
 {
     auto geometry_data = std::make_unique<SSL_GeometryData>();
 
@@ -18,7 +18,7 @@ TEST(SSLWrapperMessageTranslatorTest, test_create_wrapper_only_with_geometry)
     EXPECT_FALSE(wrapper_packet->has_detection());
 }
 
-TEST(SSLWrapperMessageTranslatorTest, test_create_wrapper_only_with_detection_frame)
+TEST(SSLWrapperTest, test_create_wrapper_only_with_detection_frame)
 {
     auto detection_frame = std::make_unique<SSL_DetectionFrame>();
 
@@ -27,7 +27,7 @@ TEST(SSLWrapperMessageTranslatorTest, test_create_wrapper_only_with_detection_fr
     EXPECT_TRUE(wrapper_packet->has_detection());
 }
 
-TEST(SSLWrapperMessageTranslatorTest, test_create_wrapper_with_all_data)
+TEST(SSLWrapperTest, test_create_wrapper_with_all_data)
 {
     auto geometry_data   = std::make_unique<SSL_GeometryData>();
     auto detection_frame = std::make_unique<SSL_DetectionFrame>();
