+++ src/software/sensor_fusion/BUILD
@@ -42,7 +42,9 @@ cc_library(
         "//software/parameter:dynamic_parameters",
         "//software/primitive",
         "//software/proto:sensor_msg_cc_proto",
-        "//software/proto/message_translation:ssl_protobuf_reader",
+        "//software/proto/message_translation:ssl_detection",
+        "//software/proto/message_translation:ssl_geometry",
+        "//software/proto/message_translation:ssl_referee",
         "//software/sensor_fusion/filter:sensor_fusion_filters",
         "//software/world",
     ],
